+++ src/libs/antares/exception/LoadingError.cpp
@@ -62,9 +62,25 @@ InvalidStudy::InvalidStudy(const Yuni::String& study) :
 {
 }
 
-InvalidVersion::InvalidVersion(const char* version, const char* latest) :
- LoadingError(std::string("Invalid version for the study : found `") + version + "`, expected <=`"
-              + latest + '`')
+static std::string InvalidSolverSpecificParametersHelper(const std::string& solver,
+                                                         const std::string& specificParameters)
+{
+    std::ostringstream message;
+    message << "Specific parameters '" << specificParameters
+            << "' are not valid or not supported for solver " << solver;
+    return message.str();
+}
+
+InvalidSolverSpecificParameters::InvalidSolverSpecificParameters(
+  const std::string& solver,
+  const std::string& specificParameters):
+    LoadingError(InvalidSolverSpecificParametersHelper(solver, specificParameters))
+{
+}
+
+InvalidStudy::InvalidStudy(const Yuni::String& study):
+    LoadingError(std::string("The folder `") + study.c_str()
+                 + "` does not seem to be a valid study")
 {
 }