+++ src/tests/end-to-end/simple_study/simple-study.cpp
@@ -151,8 +151,8 @@ BOOST_AUTO_TEST_CASE(milp_two_mc_single_unit_single_scenario)
     // Use OR-Tools / COIN for MILP
     auto& p = study->parameters;
     p.unitCommitment.ucMode = ucMILP;
-    p.ortoolsUsed = true;
-    p.ortoolsSolver = "coin";
+    p.optOptions.ortoolsUsed = true;
+    p.optOptions.ortoolsSolver = "coin";
 
     simulation->create();
     simulation->run();
@@ -180,8 +180,8 @@ BOOST_AUTO_TEST_CASE(milp_two_mc_two_unit_single_scenario)
     // Use OR-Tools / COIN for MILP
     auto& p = study->parameters;
     p.unitCommitment.ucMode = ucMILP;
-    p.ortoolsUsed = true;
-    p.ortoolsSolver = "coin";
+    p.optOptions.ortoolsUsed = true;
+    p.optOptions.ortoolsSolver = "coin";
 
     simulation->create();
     simulation->run();